refactor(editor): studio controller 파일 분리
